using PermissionUtils methods, plus added optimization to allow to stay on the Activity while permissions granting process is ongoing

+++ stories/src/main/java/com/wordpress/stories/compose/ComposeLoopFrameActivity.kt
@@ -198,6 +198,7 @@ abstract class ComposeLoopFrameActivity : AppCompatActivity(), OnStoryFrameSelec
     private var storyDiscardListener: StoryDiscardListener? = null
     private var notificationTrackerProvider: NotificationTrackerProvider? = null
     private var firstIntentLoaded: Boolean = false
+    private var permissionsRequestInProcess: Boolean = false
 
     private val connection = object : ServiceConnection {
         override fun onServiceConnected(className: ComponentName, service: IBinder) {
@@ -463,7 +464,9 @@ abstract class ComposeLoopFrameActivity : AppCompatActivity(), OnStoryFrameSelec
     private fun setupStoryViewModelObservers() {
         storyViewModel.uiState.observe(this, Observer {
             // if no frames in Story, finish
-            if (storyViewModel.getCurrentStorySize() == 0 && firstIntentLoaded) {
+            // note momentarily there will be times when this LiveData is triggered while permissions are
+            // being requested so, don't proceed if that is the case
+            if (storyViewModel.getCurrentStorySize() == 0 && firstIntentLoaded && !permissionsRequestInProcess) {
                 // finally, delete the captured media
                 deleteCapturedMedia()
                 finish()
@@ -649,6 +652,7 @@ abstract class ComposeLoopFrameActivity : AppCompatActivity(), OnStoryFrameSelec
     override fun onRequestPermissionsResult(requestCode: Int, permissions: Array<String>, grantResults: IntArray) {
         if (PermissionUtils.allRequiredPermissionsGranted(this)) {
             onLoadFromIntent(intent)
+            permissionsRequestInProcess = false
         } else {
             super.onRequestPermissionsResult(requestCode, permissions, grantResults)
         }
@@ -698,7 +702,7 @@ abstract class ComposeLoopFrameActivity : AppCompatActivity(), OnStoryFrameSelec
                     addFramesToStoryFromMediaUriList(uriList)
                     setDefaultSelectionAndUpdateBackgroundSurfaceUI()
                 } else if (data.hasExtra(requestCodes.EXTRA_LAUNCH_WPSTORIES_CAMERA_REQUESTED)) {
-                    if (!areAllNeededPermissionsGranted()) {
+                    if (!PermissionUtils.allRequiredPermissionsGranted(this)) {
                         // at this point, the user wants to launch the camera
                         // but we need to check whether we have permissions for thatt.
                         // after permissions are requestted, we need the original intent to be set differently
@@ -1085,20 +1089,10 @@ abstract class ComposeLoopFrameActivity : AppCompatActivity(), OnStoryFrameSelec
         photoEditor.addNewImageView(true, Uri.parse("https://i.giphy.com/Ok4HaWlYrewuY.gif"))
     }
 
-    private fun areAllNeededPermissionsGranted(): Boolean {
-        return (!PermissionUtils.checkPermission(this, Manifest.permission.RECORD_AUDIO) ||
-                !PermissionUtils.checkPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) ||
-                !PermissionUtils.checkPermission(this, Manifest.permission.CAMERA))
-    }
-
     private fun launchCameraPreview() {
-        if (!areAllNeededPermissionsGranted()) {
-            val permissions = arrayOf(
-                Manifest.permission.RECORD_AUDIO,
-                Manifest.permission.WRITE_EXTERNAL_STORAGE,
-                Manifest.permission.CAMERA
-            )
-            PermissionUtils.requestPermissions(this, permissions)
+        if (!PermissionUtils.allRequiredPermissionsGranted(this)) {
+            permissionsRequestInProcess = true
+            PermissionUtils.requestAllRequiredPermissions(this)
             return
         }
